Revert "Vendor go-kext and go-keychain (#10795)"
This reverts commit ed1b6584cb1207d98171375e94c57278c635da99.
