+++ go/pinentry/pinentry_osx.go
@@ -10,6 +10,8 @@ import (
 	"encoding/hex"
 	"fmt"
 	"os"
+
+	"github.com/keybase/go-keychain"
 )
 
 const (
@@ -31,7 +33,8 @@ func (pi *pinentryInstance) useSecretStore(useSecretStore bool) (pinentrySecretS
 
 	// Make account name unpredictable to make it infeasible for
 	// an attacker to guess (and thus sniff the passphrase). See
-	// https://github.com/keybase/client/issues/484#issuecomment-114313867 .
+	// https://github.com/keybase/client/issues/484#issuecomment-114313867
+	// .
 	var accountNameBytes [accountNameByteLength]byte
 	n, err := rand.Read(accountNameBytes[:])
 	if n != accountNameByteLength {
@@ -58,6 +61,53 @@ func (pi *pinentryInstance) useSecretStore(useSecretStore bool) (pinentrySecretS
 		return "", nil
 	}
 	return pinentrySecretStoreInfo(accountName), err
+}
+
+func (pi *pinentryInstance) shouldStoreSecret(info pinentrySecretStoreInfo) bool {
+	if len(info) == 0 {
+		return false
+	}
+
+	// We just want to know when the user did check the "Save in
+	// Keychain" checkbox, so remove whatever pinentry put into
+	// the keychain, and infer the state of the checkbox from the
+	// error (since there will be no error if an entry was found
+	// and deleted).
+	//
+	// This is a bit of a hack -- this may cause a dialog to pop
+	// up saying that the client wants to access the user's
+	// keychain. But this will do for now until we write our own
+	// pinentry.
+	query := keychain.NewItem()
+	query.SetSecClass(keychain.SecClassGenericPassword)
+	query.SetService(pinentryServiceName)
+	query.SetAccount(string(info))
+	query.SetMatchLimit(keychain.MatchLimitOne)
+
+	// We need to query and delete by item reference because the
+	// OSX keychain API only allows us to delete unowned items
+	// this way.
+	query.SetReturnRef(true)
+	ref, err := keychain.QueryItemRef(query)
+	if err != nil {
+		// Default to false if there was an error.
+		return false
+	}
+	if ref == nil {
+		// If not found, return false.
+		return false
+	}
+
+	defer keychain.Release(ref)
+
+	err = keychain.DeleteItemRef(ref)
+	if err != nil {
+		// Default to false if there was an error deleting.
+		return false
+	}
+
+	// Entry was found and deleted.
+	return true
 }
 
 func HasWindows() bool {
